% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_albums.R
\name{get_albums}
\alias{get_albums}
\title{Get JSON file of albums}
\usage{
get_albums(x, token)
}
\arguments{
\item{x}{a spotify album id, e.g. as returned by get_tracks()}

\item{token}{token returned by spotifyOAuth()}
}
\value{
JSON
}
\description{
Get JSON file of albums
}
\examples{
# get_albums(saved_tracks$album_id, token = my_token)
}
