% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_artists.R
\name{get_artists}
\alias{get_artists}
\title{Get JSON file of artists}
\usage{
get_artists(artists, by = "name", token)
}
\arguments{
\item{artists}{either spotify artist_ids or strings of artist names (if given actual names, a search will be done)}

\item{by}{can be "name" or "id"}

\item{token}{token returned by spotifyOAuth()}
}
\value{
JSON
}
\description{
blabla
}
\examples{
# get_artists("James Last Orchestra", token = my_token)

# x <- c("James Last", "Britney Spears", "Metallica")
# get_artists(x, token = my_token)
}
