% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_artists_album.R
\name{get_artists_albums}
\alias{get_artists_albums}
\title{Get JSON file of albums}
\usage{
get_artists_albums(x, by = "artist_id", n = 20, token)
}
\arguments{
\item{x}{a spotify id, e.g. as returned by get_artists()}

\item{by}{either "artist_id" or "album_id"; DEFUNCT FOR NOW}

\item{n}{number of albums to return per \code{artist_id} (if \code{n} is bigger than number available albums, all albums will be returned)}

\item{token}{token returned by spotifyOAuth()}
}
\value{
JSON
}
\description{
Get JSON file of albums
}
\examples{
# get_artists_albums(tidy_artists$album_id, n = 5, token = my_token)
}
