% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_get_playlist.R
\name{get_playlist}
\alias{get_playlist}
\title{Get JSON file of a playlist's tracks}
\usage{
get_playlist(playlist_id, token)
}
\arguments{
\item{playlist_id}{a spotify id, e.g. as returned by get_songs_from_albums()}

\item{token}{token returned by spotifyOAuth()}
}
\value{
JSON
}
\description{
Get JSON file of a playlist's tracks
}
\examples{
# get_song_features(tidy_tracks$track_id, my_token)
}
