% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_track_features.R
\name{get_track_features}
\alias{get_track_features}
\title{Get JSON file of song features}
\usage{
get_track_features(track_id, token)
}
\arguments{
\item{track_id}{a spotify id, e.g. as returned by get_songs_from_albums()}

\item{token}{token returned by spotifyOAuth()}
}
\value{
JSON
}
\description{
Get JSON file of song features
}
\examples{
# get_song_features(tidy_tracks$track_id, my_token)
}
