% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tracks.R
\name{get_tracks}
\alias{get_tracks}
\title{Get JSON file of an album's songs}
\usage{
get_tracks(track_ids, token)
}
\arguments{
\item{track_ids}{a spotify id, e.g. as returned by get_artists_album()}

\item{token}{token returned by spotifyOAuth()}
}
\value{
JSON
}
\description{
Get JSON file of an album's songs
}
\examples{
# get_songs_from_albums(tidy_albums$album_id, my_token)
}
