% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_saved.R
\name{get_user_saved}
\alias{get_user_saved}
\title{Get JSON file of current users saved tracks or albums}
\usage{
get_user_saved(type = "albums", n = 20, token)
}
\arguments{
\item{type}{can be "albums" (default) or "tracks"}

\item{n}{number of entries to return}

\item{token}{token returned by spotifyOAuth()}
}
\value{
JSON
}
\description{
Get JSON file of current users saved tracks or albums
}
\examples{
# nothing to see for now
}
