% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_toplist.R
\name{get_user_toplist}
\alias{get_user_toplist}
\title{Get JSON file of current users top tracks or artists}
\usage{
get_user_toplist(type = "artists", time_range = "medium_term", n = 20, token)
}
\arguments{
\item{type}{can be "artists" or "tracks"}

\item{time_range}{can be "short_term" (last ~4 weeks), "medium_term" (last ~6 months) or "long_term" (several years)}

\item{n}{number of entries to return (must be between 1 and 50)}

\item{token}{token returned by spotifyOAuth()}
}
\value{
JSON
}
\description{
Get JSON file of current users top tracks or artists
}
\examples{
# get_user_toplist(token = my_token) # get users top 20 artists of the last 6 months
# get_user_toplist("tracks", "short_term", 10, my_token) # get users top 10 songs of the last 4 weeks
}
