% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OAuth.R
\name{spotifyOAuth}
\alias{spotifyOAuth}
\title{Get an OAuth token from spotify}
\usage{
spotifyOAuth(app_id, client_id, client_secret)
}
\arguments{
\item{app_id}{the name of your spotify app as}

\item{client_id}{your client id, semi-optimally stored in your .Renviron}

\item{client_secret}{your client secret, semi-optimally stored in your .Renviron}
}
\value{
OAuth token
}
\description{
credentials can be looked up at your spotify dev / app
client_id and client_secret are stored in my .Renviron
}
\details{
this function(s) are humbly taken in about equal parts from:
https://github.com/tiagomendesdantas/Rspotify
https://github.com/jemus42/tRakt/blob/master/R/api.R
}
\examples{
# none yet
}
