% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_artists_album.R
\name{tidy_albums}
\alias{tidy_albums}
\title{Create a tidy dataframe of albums from JSON}
\usage{
tidy_albums(albums_json, image_size = 2)
}
\arguments{
\item{albums_json}{a JSON object of albums, e.g. as returned by get_artists_albums()}

\item{image_size}{1, 2, 3 /shrug}
}
\value{
dataframe
}
\description{
Create a tidy dataframe of albums from JSON
}
\examples{
# tidy_albums(albums_json)
}
