% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_artists.R
\name{tidy_artists}
\alias{tidy_artists}
\title{Create a tidy dataframe of artists from JSON}
\usage{
tidy_artists(artists_json, image_size = 2)
}
\arguments{
\item{artists_json}{a JSON object of artists, e.g. as returned by get_artists()}

\item{image_size}{1 (640x640 pixel image), 2 (320x320) or 3 (160x160)}
}
\value{
dataframe
}
\description{
Create a tidy dataframe of artists from JSON
}
\examples{
# tidy_artists(artists_json, 1)
}
