% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_get_playlist.R
\name{tidy_playlist}
\alias{tidy_playlist}
\title{Create a tidy dataframe from song informations and features from JSON and df}
\usage{
tidy_playlist(playlist_json)
}
\arguments{
\item{playlist_json}{a JSON object of playlist tracks, e.g. as returned by get_playlist()}
}
\value{
dataframe
}
\description{
Create a tidy dataframe from song informations and features from JSON and df
}
\examples{
# tidy_track_features(playlist_json, tidy_tracks)
}
