% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tracks.R
\name{tidy_tracks}
\alias{tidy_tracks}
\title{Create a tidy dataframe songs from JSON}
\usage{
tidy_tracks(tracks_json)
}
\arguments{
\item{tracks_json}{a JSON object of songs, e.g. as returned by get_songs_from_albums()}
}
\value{
dataframe
}
\description{
Create a tidy dataframe songs from JSON
}
\examples{
# tidy_songs_from_albums(songs_json)
}
